+++ towr/src/total_duration_constraint.cc
@@ -27,43 +27,48 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ******************************************************************************/
 
-#include <towr/costs/soft_constraint.h>
+#include <towr/constraints/total_duration_constraint.h>
+#include <towr/variables/variable_names.h>
 
 namespace towr {
 
-SoftConstraint::SoftConstraint (const ConstraintPtr& constraint)
-    :Component(1, "SoftConstraint-" + constraint->GetName())
+
+TotalDurationConstraint::TotalDurationConstraint (double T_total, int ee)
+    :ConstraintSet(1, "totalduration-" + std::to_string(ee))
 {
-  constraint_ = constraint;
-  int n_constraints = constraint_->GetRows();
-
-  // average value of each upper and lower bound
-  b_ = VectorXd(n_constraints);
-  int i=0;
-  for (auto b : constraint_->GetBounds()) {
-    b_(i++) = (b.upper_ + b.lower_)/2.;
-  }
-
-  // treat all constraints equally by default
-  W_.resize(n_constraints);
-  W_.setOnes();
+  T_total_ = T_total;
+  ee_ = ee;
 }
 
-SoftConstraint::VectorXd
-SoftConstraint::GetValues () const
+void
+TotalDurationConstraint::InitVariableDependedQuantities (const VariablesPtr& x)
 {
-  VectorXd g = constraint_->GetValues();
-  VectorXd cost = 0.5*(g-b_).transpose()*W_.asDiagonal()*(g-b_);
-  return cost;
+  phase_durations_ = x->GetComponent<PhaseDurations>(id::EESchedule(ee_));
 }
 
-SoftConstraint::Jacobian
-SoftConstraint::GetJacobian () const
+Eigen::VectorXd
+TotalDurationConstraint::GetValues () const
 {
-  VectorXd g   = constraint_->GetValues();
-  Jacobian jac = constraint_->GetJacobian();
-  VectorXd grad = jac.transpose()*W_.asDiagonal()*(g-b_);
-  return grad.transpose().sparseView();
+  VectorXd g = VectorXd::Zero(GetRows());
+  g(0) = phase_durations_->GetValues().sum(); // attention: excludes last duration
+  return g;
 }
 
-} /* namespace towr */
+TotalDurationConstraint::VecBound
+TotalDurationConstraint::GetBounds () const
+{
+  // TODO hacky and should be fixed
+  // since last phase is not optimized over these hardcoded numbers go here
+  int min_duration_last_phase = 0.2;
+  return VecBound(GetRows(), ifopt::Bounds(0.1, T_total_-min_duration_last_phase));
+}
+
+void
+TotalDurationConstraint::FillJacobianBlock (std::string var_set, Jacobian& jac) const
+{
+  if (var_set == phase_durations_->GetName())
+    for (int col=0; col<phase_durations_->GetRows(); ++col)
+      jac.coeffRef(0, col) = 1.0;
+}
+
+} // namespace towr