+++ towr_ros/include/towr_ros/height_map_examples.h
@@ -72,9 +72,9 @@ public:
   virtual double GetHeightDerivWrtX(double x, double y) const override;
 
 private:
-  double block_start = 1.5;
+  double block_start = 0.7;
   double length_     = 3.5;
-  double height_     = 0.8; // [m]
+  double height_     = 0.5; // [m]
 
   double eps_ = 0.03; // approximate as slope
   const double slope_ = height_/eps_;
@@ -86,7 +86,7 @@ public:
   virtual double GetHeight(double x, double y) const override;
 
 private:
-  double first_step_start_  = 1.5;
+  double first_step_start_  = 1.0;
   double first_step_width_  = 0.4;
   double height_first_step  = 0.2;
   double height_second_step = 0.4;
@@ -101,12 +101,12 @@ public:
   virtual double GetHeightDerivWrtXX(double x, double y) const override;
 
 private:
-  const double gap_start_ = 1.5;
-  const double w = 1.0; // gap width or 0.5 for ANYmal
-  const double h = 1.0; // 1.6 was
+  const double gap_start_ = 1.0;
+  const double w = 0.6;
+  const double h = 1.5;
 
   const double slope_ = h/w;
-  const double dx = w/2.0; // gap witdh 2
+  const double dx = w/2.0;
   const double xc = gap_start_ + dx; // gap center
   const double gap_end_x = gap_start_ + w;
 
@@ -143,10 +143,10 @@ public:
   virtual double GetHeightDerivWrtY(double x, double y) const override;
 
 private:
-  const double x_start_ = 1.5;
+  const double x_start_ = 1.0;
   const double length_  = 1.5;
   const double y_start_ = 0.5; // distance to start of slope from center at z=0
-  const double slope_   = 3;   // 2 or 3
+  const double slope_   = 3.0;
 
   const double x_end_ = x_start_+length_;
 };