+++ towr_ros/src/towr_ros.cc
@@ -81,19 +81,26 @@ TowrRos::TowrRos ()
 
   std::vector<Eigen::Vector3d> ee_pos(4);
 
-  ee_pos.at(0) <<  0.31,  0.29, 0.0; // LF
-  ee_pos.at(1) <<  0.31, -0.29, 0.0; // RF
-  ee_pos.at(2) << -0.31,  0.29, 0.0; // LH
-  ee_pos.at(3) << -0.31, -0.29, 0.0; // RH
+  ee_pos.at(0) <<  0.31,  0.29, 0.0;
+  ee_pos.at(1) <<  0.31, -0.29, 0.0;
+  ee_pos.at(2) << -0.31,  0.29, 0.0;
+  ee_pos.at(3) << -0.31, -0.29, 0.0;
 
   towr_.SetInitialState(b, ee_pos);
 
   model_.dynamic_model_   = std::make_shared<towr::HyqDynamicModel>();
   model_.kinematic_model_ = std::make_shared<towr::HyqKinematicModel>();
   gait_                   = std::make_shared<towr::QuadrupedGaitGenerator>();
-  output_dt_              = 0.0025; // 400 Hz control loop frequency on ANYmal
+  output_dt_              = 0.02;
 
-  terrain_ = std::make_shared<FlatGround>(0.0);
+  ground_height_ = 0.0;
+  terrain_ = std::make_shared<FlatGround>(ground_height_);
+
+  // could also use SNOPT here
+  solver_ = std::make_shared<ifopt::Ipopt>();
+  solver_->max_cpu_time_ = 10.0;
+  solver_->linear_solver_ = "ma27";
+  solver_->use_jacobian_approximation_ = false;
 }
 
 void
@@ -103,13 +110,17 @@ TowrRos::CurrentStateCallback (const xpp_msgs::RobotStateCartesian& msg)
   base_initial.lin.at(towr::kPos) = xpp::Convert::ToXpp(msg.base.pose.position);
   base_initial.lin.at(towr::kVel) = xpp::Convert::ToXpp(msg.base.twist.linear);
 
-  Eigen::Quaterniond q = xpp::Convert::ToXpp(msg.base.pose.orientation);
-  base_initial.ang.at(towr::kPos) = GetUnique(xpp::GetEulerZYXAngles(q));
+  Eigen::Quaterniond q_BtoW = xpp::Convert::ToXpp(msg.base.pose.orientation);
+  base_initial.ang.at(towr::kPos) = GetUnique(xpp::GetEulerZYXAngles(q_BtoW));
   base_initial.ang.at(towr::kVel) = Vector3d::Zero(); // smell fill this angular_vel->euler rates
 
   std::vector<Vector3d> feet_initial;
-  for (auto ee : msg.ee_motion)
-    feet_initial.push_back(xpp::Convert::ToXpp(ee.pos));
+  ground_height_ = 0.0;
+  for (auto ee : msg.ee_motion) {
+    Vector3d pos = xpp::Convert::ToXpp(ee.pos);
+    feet_initial.push_back(pos);
+    ground_height_ += pos.z()/msg.ee_motion.size(); // adapt ground height based on avg initial foothold height
+  }
 
   towr_.SetInitialState(base_initial, feet_initial);
 }
@@ -174,10 +185,8 @@ TowrRos::SetTowrParameters(const TowrCommand& msg)
     params.ee_in_contact_at_start_.push_back(gait_->IsInContactAtStart(ee));
   }
 
-
-  double ground_height = 0.0; // must possibly be adapted for real robot
   auto terrain_id = static_cast<towr::TerrainID>(msg.terrain_id);
-  terrain_ = towr::HeightMapFactory::MakeTerrain(terrain_id, ground_height);
+  terrain_ = towr::HeightMapFactory::MakeTerrain(terrain_id, ground_height_);
 
 
   towr_.SetParameters(goal, params, model_, terrain_);
@@ -187,7 +196,7 @@ void
 TowrRos::OptimizeMotion ()
 {
   try {
-    towr_.SolveNLP();
+    towr_.SolveNLP(solver_);
   } catch (const std::runtime_error& e) {
     ROS_ERROR_STREAM("Optimization failed, not sending. " << e.what());
   }
@@ -426,15 +435,6 @@ TowrRos::GetUnique (const Vector3d& zyx_non_unique) const
 
   return zyx;
 }
-
-//  void SetTerrainHeightFromAvgFootholdHeight(
-//      HeightMap::Ptr& terrain) const
-//  {
-//    double avg_height=0.0;
-//    for ( auto pos : new_ee_pos_)
-//      avg_height += pos.z()/new_ee_pos_.size();
-//    terrain->SetGroundHeight(avg_height);
-//  }
 
 } /* namespace towr */